+++ tests/Unit/Public/Install-SqlDscServer.Tests.ps1
@@ -77,7 +77,7 @@ Describe 'Install-SqlDscServer' -Tag 'Public' {
         @{
             MockParameterSetName   = 'Upgrade'
             # cSpell: disable-next
-            MockExpectedParameters = '-Upgrade -AcceptLicensingTerms -MediaPath <string> -InstanceName <string> [-Enu] [-UpdateEnabled] [-UpdateSource <string>] [-InstanceDir <string>] [-InstanceId <string>] [-ProductKey <string>] [-BrowserSvcStartupType <string>] [-FTUpgradeOption <string>] [-ISSvcAccount <string>] [-ISSvcPassword <securestring>] [-ISSvcStartupType <string>] [-AllowUpgradeForSSRSSharePointMode] [-FailoverClusterRollOwnership <ushort>] [-ProductCoveredBySA] [-Timeout <uint>] [-Force] [-WhatIf] [-Confirm] [<CommonParameters>]'
+            MockExpectedParameters = '-Upgrade -AcceptLicensingTerms -MediaPath <string> -InstanceName <string> [-Enu] [-UpdateEnabled] [-UpdateSource <string>] [-InstanceDir <string>] [-InstanceId <string>] [-ProductKey <string>] [-BrowserSvcStartupType <string>] [-FTUpgradeOption <string>] [-ISSvcAccount <string>] [-ISSvcPassword <securestring>] [-ISSvcStartupType <string>] [-AllowUpgradeForSSRSSharePointMode] [-AllowDqRemoval] [-FailoverClusterRollOwnership <ushort>] [-ProductCoveredBySA] [-Timeout <uint>] [-Force] [-WhatIf] [-Confirm] [<CommonParameters>]'
         }
         @{
             MockParameterSetName   = 'EditionUpgrade'
@@ -784,6 +784,11 @@ Describe 'Install-SqlDscServer' -Tag 'Public' {
                 MockParameterName  = 'FailoverClusterRollOwnership'
                 MockParameterValue = 2
                 MockExpectedRegEx  = '\/FAILOVERCLUSTERROLLOWNERSHIP=2' # cspell: disable-line
+            }
+            @{
+                MockParameterName  = 'AllowDqRemoval'
+                MockParameterValue = $true
+                MockExpectedRegEx  = '\/IACCEPTDQUNINSTALL\s*' # cspell: disable-line
             }
         ) {
             BeforeAll {