+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `Install-SqlDscServer`
+  - Added parameter `AllowDqRemoval` to the `Upgrade` parameter set
+    ([issue #2155](https://github.com/dsccommunity/SqlServerDsc/issues/2155)).
+- `Test-SqlDscIsSupportedFeature`
+  - Added DQ, DQC, and MDS features as discontinued starting with SQL Server 2025
+    (17.x) and later versions ([issue #2380](https://github.com/dsccommunity/SqlServerDsc/issues/2380)).
 - Added public command `Get-SqlDscRSPackage` to retrieve package information for
   SQL Server Reporting Services or Power BI Report Server. Supports getting version
   information from an executable file
@@ -59,6 +65,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Added parameter `-KillActiveSessions` to automatically terminate any active
     sessions for a login before dropping it
     ([issue #2372](https://github.com/dsccommunity/SqlServerDsc/issues/2372)).
+- `Invoke-SetupAction`
+  - Added parameter `AllowDqRemoval` for the `Upgrade` action to allow removal
+    of Data Quality (DQ) Services during upgrade to SQL Server 2025 (17.x) and
+    later versions.
+  - Now outputs setup progress when `-Verbose` is passed by using `/QUIETSIMPLE`
+    instead of `/QUIET`.
 
 ### Changed